refactor: help schema refactor
