+++ src/schemas/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod common;
 pub(crate) mod configuration_management;
 pub(crate) mod fonts;
-pub(crate) mod help_schema;
+pub(crate) mod help;
 pub(crate) mod os_settings;
 pub(crate) mod shell_configuration;
 pub(crate) mod state_file;
